+++ src/main/java/freshtrash/freshtrashbackend/controller/AuthApi.java
@@ -1,7 +1,9 @@
 package freshtrash.freshtrashbackend.controller;
 
+import freshtrash.freshtrashbackend.dto.request.LoginRequest;
 import freshtrash.freshtrashbackend.dto.request.SignUpRequest;
 import freshtrash.freshtrashbackend.dto.response.ApiResponse;
+import freshtrash.freshtrashbackend.dto.response.LoginResponse;
 import freshtrash.freshtrashbackend.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -18,9 +20,21 @@ import javax.validation.Valid;
 public class AuthApi {
     private final MemberService memberService;
 
+    /**
+     * 회원가입
+     */
     @PostMapping("/signup")
     public ResponseEntity<?> signup(@RequestBody @Valid SignUpRequest signUpRequest) {
         memberService.registerMember(signUpRequest.toEntity());
         return ResponseEntity.ok(ApiResponse.of("you're successfully sign up. you can be login."));
+    }
+
+    /**
+     * 로그인
+     */
+    @PostMapping("/signin")
+    public ResponseEntity<LoginResponse> signIn(@RequestBody @Valid LoginRequest loginRequest) {
+        LoginResponse loginResponse = memberService.signIn(loginRequest.email(), loginRequest.password());
+        return ResponseEntity.ok(loginResponse);
     }
 }
